Created secondary port interface for a Todo repository
This will be the interface that secondary adapters need to implement
in order to be compatible with our application. For now, these just
implement basic CRUD operations. Over time, we'll evolve to have
different adapters that we plug in to our application for various
scenarios such as using a message queue or database.