Updated domain model to include event sourced domain model
Preemptively refactored the secondary adapter file to a folder, as we will
eventually be updating our application with multiple adapters.

Added event interfaces to a new sub module in the application domain. These
events are the foundation of pivoting the application to an event sourced
model. The fundamental shift in the mental model of event sourcing is that
all application state is transient. The source of truth that is persisted
is the event log, which has a history of every event that has happened in the
system.

For now, we have modeled the same services we created in the beginning
for our basic TODO example as events. These events will eventually be used
to hydrate "projections" that represent our domain model for some context so
that we can model the application state in a way that satifies our requirements.