Implemented event sourced model for To-Do list
Updated the domain model to create an event sourced model of our
To-Do application. Also implemented a new repository implementation
that implements a new secondary adapter interface for interacting
with a todo event repository. These will be used when creating a
facade that interacts with the To-Do application by persisting
and updating the state based on an event log.