Refactored TodoFacade to use event repository
To implement the To-Do list logic using an event sourced approach,
updated the TodoFacade that implements our primary TodoService
adapter to use the new event based repository.

Additionally updated imports and points where we plug in the
secondary adapters during the construction in the cli adapter
to plug in the new implementation of the InMemoryEventRepository
to persist application events in memory.