Fixed 'this' context in typescript classes
Updated classes for our TodoList and TodoTask projections to use arrow
functions instead of methods on a class. Because Typescript still
ultimately compiles to Javascript, there is still very unusual behavior
regarding the context and scope of 'this' at runtime, which was
discovered to be causing bugs at runtime when testing out the CLI
adapter.

More about that here
https://www.typescriptlang.org/docs/handbook/2/classes.html#this-at-runtime-in-classes