Implemented an AWS SQS connected event repository
Added new dependencies for aws-sdk and the aws-sdk/client-sqs packages
for communicating with SQS. Used these dependencies to implement a
new secondary adapter that interacts with SQS for writing events
instead of just saving them in memory. In future work, we'll use
the messages from that queue to update projections in dynamoDB
that will be processed by workers that listen to and respond to
incoming queue messages.