Implemented SQS event publishing
Added SQS event publishing on save for the aws event repository.
Added in-memory data store for storing local copy of the data for now.
In the future, will have worker processes processing the messages and
saving the results in a proper database to be read by the app-facing
server.

Note: for some reason, I couldn't get the sqs batch message sending
working. I left the code that implements message batching so that
when I get to figure out what the issue was, messages can be batched
to SQS for cost savings (especially if there's more than one event
that resulted from a command).